opendmarc-arcares.c:arc_parse: tokens are separated by colon
And not by colon or space.  The latter cannot deal with
“arc=reject (some comment)”.  On the first iteration of the loop,
tag_code is AAR_ARC_TAG_ARC and tag_value is reject.  On the second
iteration, there is no equal sign in token_ptr, so token_ptr is
NULL and then opendmarc_arcares_strip_whitespace(NULL) is called,
which does not work.

opendmarc uses anyway only the smtpclienip from the parsed AAR.