libopendmarc: fix a critical bug in opendmarc_policy_store_dkim
in case multiple DKIM signatures are present in a message, a certain
order (first, invalid DKIM signature for aligned domain, then valid
signature for arbitrary invalid domain)
of storing the corresponding DKIM results can lead to a DKIM
signature from one domain be accepted as PASS for another domain. This
was possible because a path existed to overwrite a non-PASS DKIM result
*without* updating the domain name. To prevent this, a condition is
added that makes sure to always update the domain name if the DKIM
result is updated.