Fix detection of res_ndestroy
The author of 8821364d1bb49069ae332d1529b1dc95a36131d8 did not
fully understand the code they were modifying, so the effect of their
change was to define `HAVE_RES_NDESTROY` if the function `res_ninit`
existed, rather than if the function `res_ndestroy` existed. This
commit fixes the logic so that both `res_ninit` and `res_ndestroy` are
checked for the way `res_ninit` was before the broken commit.