Convert strings to all be "char *".
The C Library routines all treat strings as "char *".  Much of the
code here treats them as "unsigned char *" (sometimes by using
"u_char *") which results in the compiler flagging every single
instance as an incompatable pointer assignment.

This patch converts them all to be "char *" following the C
convention.  I chose this approach because it means all of the
string library functions work without casting.

I've reviewed as best I can this does not introduce bugs, but
what's really needed is a test of QA tests to confirm.